New icon, page button location
New extension icon, original and simple. Blue might not a strong color for it but I can change it later.

The button that shows the page contents window has been integrated into the webpage nav container instead of as an awkwardly floating button. An icon is shown next to it so it catches the eye easily.